+++ tests/test_processor_manager.py
@@ -1,22 +1,54 @@
 import unittest
+from tests.base_test import BaseTest
+from avista_sensors.processor_manager import ProcessorManager
+from avista_sensors.manager_state import ManagerState
+from avista_data.sensor import Sensor
+from avista_data import db
+from avista_data.unit import Unit
+import time
 
 
-class ProcessorManagerTest(unittest.TestCase):
+class ProcessorManagerTest(BaseTest):
+
+    def setUp(self):
+        super().setUp()
+        sensor = Sensor(name="Test", quantity="Quantity", module="avista_sensors.impl.random_processor",
+                       cls="RandomProcessor", unit=Unit.F)
+        db.session.add(sensor)
+        db.session.commit()
+        self.fixture = ProcessorManager(self.app)
 
     def test_start(self):
-        pass
+        self.fixture.init()
+        self.fixture.start()
+        time.sleep(5)
+        self.fixture.stop()
+        self.fixture.join()
+        self.assertEqual(ManagerState.IDLE, self.fixture.state)
 
     def test_stop(self):
-        pass
+        self.assertEqual(ManagerState.IDLE, self.fixture.state)
+        self.fixture.init()
+        self.fixture.start()
+        time.sleep(2)
+        self.assertEqual(ManagerState.EXECUTING, self.fixture.state)
+        self.fixture.stop()
+        self.fixture.join()
+        self.assertEqual(ManagerState.IDLE, self.fixture.state)
 
     def test_restart(self):
-        pass
+        self.assertEqual(ManagerState.IDLE, self.fixture.state)
+        self.fixture.restart()
+        time.sleep(2)
+        self.assertEqual(ManagerState.EXECUTING, self.fixture.state)
+        self.fixture.stop()
+        self.fixture.join()
+        self.assertEqual(ManagerState.IDLE, self.fixture.state)
 
     def test_init(self):
-        pass
-
-    def test_execute(self):
-        pass
+        self.fixture.init()
+        self.assertEqual(1, len(self.fixture.processors))
+        self.assertEqual(ManagerState.INITIALIZING, self.fixture.state)
 
 
 if __name__ == '__main__':