+++ CHANGELOG.md
@@ -11,7 +11,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deleted
 
-## [0.1.0] - 2020-12-03
+## [0.1.1](https://github.com/isu-avista/sensors/releases/tag/v0.1.1) - 2020-12-05
+### Added
+- Tests for all classes to reach 100% coverage (excluding SensorProcessor which is at 97%)
+- Fully documented all classes
+
+### Changed
+- Improved implementation of both RandomProcessor and SimulatedProcessor so that they work correctly
+- Added multithreading to the processor manager
+- Updated the [README.md](README.md)
+
+### Deleted
+
+## [0.1.0](https://github.com/isu-avista/sensors/releases/tag/v0.1.0) - 2020-12-03
 
 ### Added
 - All the base classes required